requirements.txt
asgiref==3.5.2
botocore==1.27.66
Django>=3.0,<4.0
django-storages==1.13.1
djangorestframework==3.13.1
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
psycopg2>=2.8
gunicorn==20.1.0
# social login
django-allauth==0.44.0
oauthlib==3.1.0
dj-rest-auth==2.1.4
django-redis
djangorestframework-simplejwt==4.6.0
# S3
Pillow
boto3==1.24.66